Revert "Speedup CLI loading by 30-50% where it matters (#14900)"
This reverts commit 3311a80576a939c63b31c07d7d5cd5b186a1a5fe.
